Break out ScoreDecoderTests
